+++ nzshm_model/model_versions/nshm_v1_0_4.py
@@ -22,6 +22,5 @@ def model_args_factory() -> Dict[str, Any]:
         title="NSHM version 1.0.4, corrected fault geometry",
         slt_json="nshm_v1.0.4_v2.json",
         gmm_json="gmcm_nshm_v1.0.4.json",
-        gmm_xml="NZ_NSHM_GMM_LT_final_EE_new_names.xml",
         hazard_config=OpenquakeConfig(DEFAULT_HAZARD_CONFIG),
     )